Conflicts:
	www-root/core/includes/functions.inc.php
	www-root/templates/default/css/style.css

+++ www-root/core/includes/functions.inc.php
@@ -5775,7 +5775,7 @@ function communities_galleries_process_photo($original_file, $photo_id = 0) {
 		/**
 		 * Check if the original_file needs to be resized or not.
 		 */
-		if(($original_file_width > $VALID_MAX_DIMENSIONS["photo"]) || ($original_file_height > $VALID_MAX_DIMENSIONS["photo"])) {
+		if(!DEMO_MODE && (($original_file_width > $VALID_MAX_DIMENSIONS["photo"]) || ($original_file_height > $VALID_MAX_DIMENSIONS["photo"]))) {
 			switch($original_file_details["mime"]) {
 				case "image/pjpeg":
 				case "image/jpeg":
@@ -5853,17 +5853,32 @@ function communities_galleries_process_photo($original_file, $photo_id = 0) {
 				return false;
 			}
 		} else {
-			if(@move_uploaded_file($original_file, $new_file)) {
-				@chmod($new_file, 0644);
+			if (!DEMO_MODE) {
+				if(@move_uploaded_file($original_file, $new_file)) {
+					@chmod($new_file, 0644);
 
-				/**
-				 * Create the new width / height so we can use the same variables
-				 * below for thumbnail generation.
-				 */
-				$new_file_width		= $original_file_width;
-				$new_file_height	= $original_file_height;
+					/**
+					 * Create the new width / height so we can use the same variables
+					 * below for thumbnail generation.
+					 */
+					$new_file_width		= $original_file_width;
+					$new_file_height	= $original_file_height;
+				} else {
+					return false;
+				}
 			} else {
-				return false;
+				if(@copy(DEMO_PHOTO, $new_file)) {
+					@chmod($new_file, 0644);
+
+					/**
+					 * Create the new width / height so we can use the same variables
+					 * below for thumbnail generation.
+					 */
+					$new_file_width		= $original_file_width;
+					$new_file_height	= $original_file_height;
+				} else {
+					return false;
+				}
 			}
 		}
 
@@ -5984,9 +5999,20 @@ function assignments_process_file($original_file, $afversion_id = 0) {
 	}
 
 	if(!@file_exists($new_file = FILE_STORAGE_PATH."/A".$afversion_id)) {
-		if(@move_uploaded_file($original_file, $new_file)) {
-			@chmod($new_file, 0644);
-			return true;
+		if (!DEMO_MODE) {
+			if(@move_uploaded_file($original_file, $new_file)) {
+				@chmod($new_file, 0644);
+				return true;
+			} else {
+				return false;
+			}
+		} else {
+			if (copy(DEMO_ASSIGNMENT, $new_file)) {
+				@chmod($new_file, 0644);
+				return true;
+			} else {
+				return false;
+			}
 		}
 	}
 
@@ -6016,9 +6042,16 @@ function communities_shares_process_file($original_file, $csfversion_id = 0) {
 	}
 
 	if(!@file_exists($new_file = COMMUNITY_STORAGE_DOCUMENTS."/".$csfversion_id)) {
-		if(@move_uploaded_file($original_file, $new_file)) {
-			@chmod($new_file, 0644);
-			return true;
+		if (!DEMO_MODE) {
+			if(@move_uploaded_file($original_file, $new_file)) {
+				@chmod($new_file, 0644);
+				return true;
+			}
+		} else {
+			if(@copy(DEMO_FILE, $new_file)) {
+				@chmod($new_file, 0644);
+				return true;
+			}
 		}
 	}
 
@@ -7081,7 +7114,7 @@ function process_user_photo($original_file, $photo_id = 0) {
 		/**
 		 * Check if the original_file needs to be resized or not.
 		 */
-		if(($original_file_width > $VALID_MAX_DIMENSIONS["photo-width"]) || ($original_file_height > $VALID_MAX_DIMENSIONS["photo-height"])) {
+		if(!DEMO_MODE && (($original_file_width > $VALID_MAX_DIMENSIONS["photo-width"]) || ($original_file_height > $VALID_MAX_DIMENSIONS["photo-height"]))) {
 			switch($original_file_details["mime"]) {
 				case "image/pjpeg":
 				case "image/jpeg":
@@ -7158,17 +7191,32 @@ function process_user_photo($original_file, $photo_id = 0) {
 				return false;
 			}
 		} else {
-			if(@move_uploaded_file($original_file, $new_file)) {
-				@chmod($new_file, 0644);
+			if (!DEMO_MODE) {
+				if(@move_uploaded_file($original_file, $new_file)) {
+					@chmod($new_file, 0644);
 
-				/**
-				 * Create the new width / height so we can use the same variables
-				 * below for thumbnail generation.
-				 */
-				$new_file_width		= $original_file_width;
-				$new_file_height	= $original_file_height;
+					/**
+					 * Create the new width / height so we can use the same variables
+					 * below for thumbnail generation.
+					 */
+					$new_file_width		= $original_file_width;
+					$new_file_height	= $original_file_height;
+				} else {
+					return false;
+				}
 			} else {
-				return false;
+				if(@copy(DEMO_PHOTO, $new_file)) {
+					@chmod($new_file, 0644);
+
+					/**
+					 * Create the new width / height so we can use the same variables
+					 * below for thumbnail generation.
+					 */
+					$new_file_width		= $original_file_width;
+					$new_file_height	= $original_file_height;
+				} else {
+					return false;
+				}
 			}
 		}
 
@@ -16819,6 +16867,9 @@ function moveImage($source, $id, $coords, $dimensions, $type = "user", $sizes = 
 	$dimensions = explode(",", $dimensions);
 
 	if ($id) {
+		if (DEMO_MODE) {
+			$source = DEMO_PHOTO;
+		}
 
 		$image_details = getimagesize($source);
 
@@ -16837,35 +16888,34 @@ function moveImage($source, $id, $coords, $dimensions, $type = "user", $sizes = 
 			break;
 		}
 
-		if ($image) {
+        if (!DEMO_MODE) {
+            if ($image) {
+                copy($source, STORAGE_USER_PHOTOS . "/" . $id . "-upload-original");
 
-			copy($source, STORAGE_USER_PHOTOS . "/" . $id . "-upload-original");
+                $image_scale = round($image_details[0] / $dimensions[0], 2);
 
-			$image_scale = round($image_details[0] / $dimensions[0], 2);
+                foreach ($coords as $coord) {
+                    $scaled_coords[] = (int) round($coord * $image_scale, 2);
+                }
 
-			foreach ($coords as $coord) {
-				$scaled_coords[] = (int) round($coord * $image_scale, 2);
-			}
+                $path = STORAGE_USER_PHOTOS . '/../public/images/' . ($type == "team" ? "teams/" : "") . $id .'/' . $id;
 
-			$path = STORAGE_USER_PHOTOS . '/../public/images/' . ($type == "team" ? "teams/" : "") . $id .'/' . $id;
-
-			foreach ($sizes as $size_name => $size) {
-				$resized_image = imagecreatetruecolor($size["width"], $size["height"]);
-				imagecopyresampled($resized_image, $image, 0, 0, $scaled_coords[0], $scaled_coords[1], $size["width"], $size["height"], $scaled_coords[2] - $scaled_coords[0], $scaled_coords[3] - $scaled_coords[1]);
-				$scaled_image =  CACHE_DIRECTORY . "/profile-img-" . $id . "-".$size["width"]."x".$size["height"].".png";
-				imagepng($resized_image, $scaled_image);
-				if (!copy($scaled_image, STORAGE_USER_PHOTOS . "/" . $id . "-" . $size_name)) {
-					$return = false;
-				} else {
-					unlink($scaled_image);
-				}
-			}
-
-		}
+                foreach ($sizes as $size_name => $size) {
+                    $resized_image = imagecreatetruecolor($size["width"], $size["height"]);
+                    imagecopyresampled($resized_image, $image, 0, 0, $scaled_coords[0], $scaled_coords[1], $size["width"], $size["height"], $scaled_coords[2] - $scaled_coords[0], $scaled_coords[3] - $scaled_coords[1]);
+                    $scaled_image =  CACHE_DIRECTORY . "/profile-img-" . $id . "-".$size["width"]."x".$size["height"].".png";
+                    imagepng($resized_image, $scaled_image);
+                    if (!copy($scaled_image, STORAGE_USER_PHOTOS . "/" . $id . "-" . $size_name)) {
+                        $return = false;
+                    } else {
+                        unlink($scaled_image);
+                    }
+                }
+            }
+        } else {
+			copy(DEMO_PHOTO, STORAGE_USER_PHOTOS . "/" . $id . "-upload");
+        }
 	}
 
-
-		return filesize(STORAGE_USER_PHOTOS . "/" . $id . "-upload");
-
-
+	return filesize(STORAGE_USER_PHOTOS . "/" . $id . "-upload");
 }
